Prevent update or delete on non existent relay address

+++ emails/views.py
@@ -97,12 +97,13 @@ def _get_relay_address_from_id(request_data, user_profile):
         )
         return relay_address
     except RelayAddress.DoesNotExist as e:
-        print(e)
         return HttpResponse("Address does not exist")
 
 
 def _index_PUT(request_data, user_profile):
     relay_address = _get_relay_address_from_id(request_data, user_profile)
+    if not isinstance(relay_address, RelayAddress):
+        return relay_address
     if request_data.get('enabled') == 'Disable':
         # TODO?: create a soft bounce receipt rule for the address?
         relay_address.enabled = False
@@ -117,8 +118,9 @@ def _index_PUT(request_data, user_profile):
 
 def _index_DELETE(request_data, user_profile):
     relay_address = _get_relay_address_from_id(request_data, user_profile)
-    # TODO?: create hard bounce receipt rule for the address
-    relay_address.delete()
+    if isinstance(relay_address, RelayAddress):
+        # TODO?: create hard bounce receipt rule for the address
+        relay_address.delete()
     return redirect('profile')
